fix(php): guard social profiles against non-array input and fix template class prefixes
- Guard all three render_social_profiles() call sites in
  Author_Block_Base with is_array() check — prevents fatal TypeError
  when $author['social'] is a string (e.g. a URL) rather than an array
- Cast authorId to int explicitly in Author_Profile_Block::render_callback
- Rename outer list wrapper class apbl-authors-list-preview → apbl-author-list
  and inner ul apbl-authors-list → apbl-author-list-items in list.php to
  match CSS selectors
- Update author-profile-content.php wrapper divs to use apbl- prefix
  (apb-author-image → apbl-author-image, apb-author-info → apbl-author-info)
  completing the partial prefix migration started earlier

+++ includes/Blocks/Author_Profile_Block.php
@@ -60,7 +60,7 @@ class Author_Profile_Block extends Author_Block_Base {
 	 * @return string Rendered block output.
 	 */
 	public function render_callback( array $attributes, string $content, $block ): string {
-		$author_id = $attributes['authorId'] ?? 0;
+		$author_id = (int) ( $attributes['authorId'] ?? 0 );
 
 		if ( empty( $author_id ) ) {
 			return $this->render_error_message( $this->get_no_author_selected_message() );
